+++ app/src/main/java/com/tengizMKCorp/tengizExpress/ui/viewmodel/ResultByCategoryViewModel.kt
@@ -18,7 +18,7 @@ class ResultByCategoryViewModel @Inject constructor(private val getProductsByCat
         MutableStateFlow<ResponseState<ProductsByCategoryID>>(ResponseState.Empty()) //mutable state flow
     val myState: StateFlow<ResponseState<ProductsByCategoryID>> = _myState //immutable state flow
 
-    fun getInfo(categoryID: Int) {
+    fun getInfo(categoryID: Long) {
         viewModelScope.launch {
             _myState.emit(ResponseState.Loading())
             val data = getProductsByCategoriesUseCase.getProductsByCategories(categoryID)
@@ -28,11 +28,13 @@ class ResultByCategoryViewModel @Inject constructor(private val getProductsByCat
         }
 
     }
+
     private var viewChangeClickCount = 0
-    fun getClickCount(): Int{
+    fun getClickCount(): Int {
         return viewChangeClickCount
     }
-    fun increaseClickCount(){
+
+    fun increaseClickCount() {
         viewChangeClickCount++
     }
 }
